Polished code: reorganised packages accordingly changed pom added spring security dependency added logging added exception handling corrected small moments
