--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/controller/payload/LoginRequestPayload.java
@@ -1,11 +1,0 @@
-package com.crmfoodestablishment.auth.controller.payload;
-
-import lombok.Getter;
-
-@Getter
-public class LoginRequestPayload {
-
-    private String email;
-
-    private String password;
-}
