--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/controller/payload/RegisterResponsePayload.java
@@ -1,11 +1,0 @@
-package com.crmfoodestablishment.auth.controller.payload;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public class RegisterResponsePayload {
-
-    private String urlOfCreatedUser;
-
-    private TokenPairResponsePayload tokenPair;
-}
