--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/controller/payload/TokenPairResponsePayload.java
@@ -1,11 +1,0 @@
-package com.crmfoodestablishment.auth.controller.payload;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public class TokenPairResponsePayload {
-
-    private String accessToken;
-
-    private String refreshToken;
-}
