--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/controller/payload/UserCreationRequestPayload.java
@@ -1,5 +1,0 @@
-package com.crmfoodestablishment.auth.controller.payload;
-
-//TODO better to be implemented by Andriy
-public class UserCreationRequestPayload {
-}
