--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/exception/InvalidTokenException.java
@@ -1,8 +1,0 @@
-package com.crmfoodestablishment.auth.exception;
-
-public class InvalidTokenException extends RuntimeException{
-
-    public InvalidTokenException(String message) {
-        super(message);
-    }
-}
