--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/exception/WrongUserCredentialsException.java
@@ -1,8 +1,0 @@
-package com.crmfoodestablishment.auth.exception;
-
-public class WrongUserCredentialsException extends RuntimeException {
-
-    public WrongUserCredentialsException(String message) {
-        super(message);
-    }
-}
