--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/service/model/AccessToken.java
@@ -1,5 +1,0 @@
-package com.crmfoodestablishment.auth.service.model;
-
-import io.jsonwebtoken.Header;
-
-public record AccessToken(Header header, AccessTokenClaims claims) {}
