--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/service/model/AccessTokenClaims.java
@@ -1,8 +1,0 @@
-package com.crmfoodestablishment.auth.service.model;
-
-import java.util.Date;
-
-public record AccessTokenClaims(
-        Date issuedAt,
-        Date expiration
-) {}
