--- user-manager-service/src/main/java/com/crmfoodestablishment/auth/service/model/RefreshToken.java
@@ -1,6 +1,0 @@
-package com.crmfoodestablishment.auth.service.model;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Header;
-
-public record RefreshToken(Header header, Claims claims) {}
