+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/ApiErrorInfo.java
@@ -1,0 +1,20 @@
+package com.crmfoodestablishment.user_auth_service;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.http.HttpStatus;
+
+@AllArgsConstructor
+@Builder
+@Setter
+@Getter
+public class ApiErrorInfo {
+
+    private String title;
+
+    private HttpStatus status;
+
+    private String description;
+}
