+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/UserAuthServiceApplication.java
@@ -1,0 +1,15 @@
+package com.crmfoodestablishment.user_auth_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+//import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+//@EnableDiscoveryClient
+public class UserAuthServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserAuthServiceApplication.class, args);
+	}
+
+}
