+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/config/AuthConfig.java
@@ -1,0 +1,16 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+//TODO implement security config
+@Configuration
+public class AuthConfig {
+
+    @Bean
+    public PasswordEncoder encoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
