+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/controller/payload/LoginRequestPayload.java
@@ -1,0 +1,17 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.controller.payload;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequestPayload {
+
+    @Email
+    private String email;
+
+    @NotBlank
+    private String password;
+}
