+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/controller/payload/UserCreationRequestPayload.java
@@ -1,0 +1,11 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.controller.payload;
+
+import lombok.Getter;
+import lombok.Setter;
+
+//TODO better to be implemented by Andriy
+//TODO write validation
+@Getter
+@Setter
+public class UserCreationRequestPayload {
+}
