+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/exception/InvalidTokenException.java
@@ -1,0 +1,8 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.exception;
+
+public class InvalidTokenException extends RuntimeException{
+
+    public InvalidTokenException(String message) {
+        super(message);
+    }
+}
