+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/exception/RegistrationException.java
@@ -1,0 +1,8 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.exception;
+
+public class RegistrationException extends RuntimeException {
+
+    public RegistrationException(String message) {
+        super(message);
+    }
+}
