+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/exception/WrongUserCredentialsException.java
@@ -1,0 +1,8 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.exception;
+
+public class WrongUserCredentialsException extends RuntimeException {
+
+    public WrongUserCredentialsException(String message) {
+        super(message);
+    }
+}
