+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/service/model/AccessToken.java
@@ -1,0 +1,5 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.service.model;
+
+import io.jsonwebtoken.Header;
+
+public record AccessToken(Header header, AccessTokenClaims claims) {}
