+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/service/model/AccessTokenClaims.java
@@ -1,0 +1,8 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.service.model;
+
+import java.util.Date;
+
+public record AccessTokenClaims(
+        Date issuedAt,
+        Date expiration
+) {}
