+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_sevice/service/model/RefreshToken.java
@@ -1,0 +1,6 @@
+package com.crmfoodestablishment.user_auth_service.auth_sevice.service.model;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Header;
+
+public record RefreshToken(Header header, Claims claims) {}
