+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/user_manager/entity/Permission.java
@@ -1,0 +1,7 @@
+package com.crmfoodestablishment.user_auth_service.user_manager.entity;
+
+public enum Permission {    
+    CONSUMER,
+    EMPLOYEE,
+    ADMIN
+}
