+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/user_manager/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package com.crmfoodestablishment.user_auth_service.user_manager.exception;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
