--- user-manager-service/src/main/java/com/crmfoodestablishment/usermanager/UserManagerApplication.java
@@ -1,15 +1,0 @@
-package com.crmfoodestablishment.usermanager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class UserManagerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserManagerApplication.class, args);
-	}
-
-}
