--- user-manager-service/src/main/java/com/crmfoodestablishment/usermanager/entity/Permission.java
@@ -1,7 +1,0 @@
-package com.crmfoodestablishment.usermanager.entity;
-
-public enum Permission {    
-    CONSUMER,
-    EMPLOYEE,
-    ADMIN
-}
