--- user-manager-service/src/main/java/com/crmfoodestablishment/usermanager/entity/UserPermission.java
@@ -1,20 +1,0 @@
-package com.crmfoodestablishment.usermanager.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@RequiredArgsConstructor
-public class UserPermission {
-
-    @EmbeddedId
-    private UserPermissionId userPermissionId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @MapsId("userId")
-    private User user;
-}
