--- user-manager-service/src/main/java/com/crmfoodestablishment/usermanager/entity/UserPermissionId.java
@@ -1,22 +1,0 @@
-package com.crmfoodestablishment.usermanager.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-import java.io.Serializable;
-
-@Embeddable
-@Getter
-@Setter
-@RequiredArgsConstructor
-public class UserPermissionId implements Serializable {
-
-    @Column(name = "user_id", nullable = false)
-    private Integer userId;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "permission", nullable = false)
-    private Permission permission;
-}
