--- user-manager-service/src/main/java/com/crmfoodestablishment/usermanager/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package com.crmfoodestablishment.usermanager.exception;
-
-public class NotFoundException extends RuntimeException {
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
