+++ user-manager-service/src/main/resources/application.properties
@@ -1,12 +1,12 @@
-spring.application.name=user-manager-service
-server.port=0
+spring.application.name=user-auth-service
+server.port=8080
 
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
 
-spring.datasource.url=jdbc:postgresql://localhost:5432/user-manager
+spring.datasource.url=jdbc:postgresql://localhost:5432/Test_authDB
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.datasource.username=postgres
-spring.datasource.password=${USERMANAGER_DATABASE_PASSWORD}
+spring.datasource.password=postgres
 
 spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.xml
 
@@ -16,10 +16,10 @@ spring.data.redis.host=localhost
 spring.data.redis.port=6379
 spring.data.redis.database=0
 
-auth.jwt.refreshToken.secretKey=someKey
-auth.jwt.refreshToken.publicKey=someKey
-auth.jwt.refreshToken.expirationTime=someValueInSecs
+auth.jwt.refreshToken.publicKey=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAk3pU/vCdSB46wLXFlKWV/eRjiPgsNI/f5FRN/sYjn8lni3YB9odElCbggPcQvkOqg7+CnT55uDxbx0gUTM1KVCEiFpUfY3V6groaSK9S7ZMN8cswh28V1wKW1E2HYgGqycnQdwcNtpuWsWwL7OKZzf+HkiK0s+bXtjlabhuKQw1i3CfNX6mKCwLp6TpSRs5cgJW7ajpw7ZKEK284mGLKQOq7jhcz5Ow5UBmW0QsBBkASHoUfJV2wjec6fdPGPhB/zVdQSU60qWAV7FlXf7ZdyzUorWdTQ/MbaXb58puQUTcyTQlnKi8UJaaWhN5UG2g5WFzlz7fiTJ0S5by5ov8Q0QIDAQAB
+auth.jwt.refreshToken.secretKey=MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCTelT+8J1IHjrAtcWUpZX95GOI+Cw0j9/kVE3+xiOfyWeLdgH2h0SUJuCA9xC+Q6qDv4KdPnm4PFvHSBRMzUpUISIWlR9jdXqCuhpIr1Ltkw3xyzCHbxXXApbUTYdiAarJydB3Bw22m5axbAvs4pnN/4eSIrSz5te2OVpuG4pDDWLcJ81fqYoLAunpOlJGzlyAlbtqOnDtkoQrbziYYspA6ruOFzPk7DlQGZbRCwEGQBIehR8lXbCN5zp908Y+EH/NV1BJTrSpYBXsWVd/tl3LNSitZ1ND8xtpdvnym5BRNzJNCWcqLxQlppaE3lQbaDlYXOXPt+JMnRLlvLmi/xDRAgMBAAECggEALxhrSSiMiYnByGWjzXbdgcEr+gMKltR1MqSrbZdXrAC0tMgAE3qZ71fE7No5OhFAxCblrpJTa3627lo5xVbvjBMyoBSJ/Ah5kYKdvyVh7aU+2D7DZhSW3F0iznRdrfILWLo+fKw/eFeSVWFEI6cvZ6XY5NV5cFPhAfupwazAx+1jx1f6p4waT/sGjiaLZpqDyuE4LZ9S3WUK+ZrNJyHGNkXzeRlvPQ1Ir2Q9wAztbSBCblTb8ufoqEKUltyhOlvvDHwG+qv+bpX4g+hYT5O1PB3ena7zIAr2+o5uT7naKwhkGX8YJJ4n1Ey3r+3RYlbvdUTl2EIMC/kKXf+Z2kvPWwKBgQC9cqpJGhWn5JFVUsollW9d/g8samcKojfUw4VlAZpVsO+LF96iG4EC+6v0u8ONSqtTlalkgD604IFoOjmdJfA2Pfwe6TH7iweaeG1ECi8++GidXfFbk0e2OQnxyuUK5Qhruo+N1MLgpfXu/OxufANh1oj1xIk1I46OVQtMz0z1EwKBgQDHSTy1/rtwDENGeJ1zfycy2KbgCgC89M6WDZNltWnEs1rBgOtJ9b3bDTIEpWCh11CqavJ9WLofU2cxdIWadKEsGAM4Sk8LeGfkuuaZdoS09OBdDdD6FOLeIL7UZisO09Mkqb01pZzhKjevW5jniPfxE225gs7F2DUZk5L5nSNzCwKBgQCTmKlIqNTkaDQ+aHDrlSAd95j0vrUm2LlsvyPHTNdkGaSXFnKyk4EQxOj0k/DHvLv8M+a0USIJo7vOa0aFw5ABgotRvdbeJjQ3Scm/L7wdcEV5DT7EXNuLAadYxP1rXoMQA1piYretdD9GAbFmZqHv8MflykgnqVsgUuup3PwVFwKBgBYajG5PO7D42zlkxWn7D3j3SMMOuPg8WvGPveqJDZswnX9a8j340L8MHqEfSrGItTzBsfWvMq2ozYhRuUPhbcL7Qtgqfcvy5esbw9n/qEZ95qjLkIM8b9FdZm0ei+YQHpoxbXgWz2i2CwkbKsF7kqydkA0EcjTFf1Y/5QxknnrzAoGBAJxhEwKqSXdm8keucO1OKTvqV/SdJjKAPZ4yhFY9NmL069oQ7ANaRXWGmO+uDg+7jCUQGgbLE/NYzqXWVI7VthiDdaav8Vz884LYpSwr2CcHjcSUeVKdkpEkHJXNeOMXsdyf2TznSHe5AGyoPLD2WxOg3t45FKpJ3kKvDzmqRcW6
+auth.jwt.refreshToken.expirationTime=10000000
 
-auth.jwt.accessToken.secretKey=someKey
-auth.jwt.accessToken.publicKey=someKey
-auth.jwt.accessToken.expirationTime=someValueInSecs
+auth.jwt.accessToken.publicKey=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAgQ2/mavp6IbfiOk2LzfybFE8Ig+bvT5tcgyJKbidMqa+36v72mM41NIZ7gQtNqcpk/EXOcIeGAmWGxBIrO6BiLxUBoIvYrtOBYtK6UbgKdUr/BAFa5W3CGnIxz0FaWhyswNkuAqAgCCwyUsLMXqle3KH8blyi5h/RaAZ0nMTmJJ+MucQwhqmDuI11uBg3rTnKwT9B+M44gePVi0CHgZWj7jZuqHOgkGUBoCDqjnMMi4cfdHhAvNTcZ9VFn3SB/9PNvUOxAyarAIHE5e3E2a/jpTjNYJ6is6BFkfcVCB2d9RPnpjpz0AudIslIw+MGLIzZ1pblYja6JVOCzC8EvXRzwIDAQAB
+auth.jwt.accessToken.secretKey=MIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQCBDb+Zq+noht+I6TYvN/JsUTwiD5u9Pm1yDIkpuJ0ypr7fq/vaYzjU0hnuBC02pymT8Rc5wh4YCZYbEEis7oGIvFQGgi9iu04Fi0rpRuAp1Sv8EAVrlbcIacjHPQVpaHKzA2S4CoCAILDJSwsxeqV7cofxuXKLmH9FoBnScxOYkn4y5xDCGqYO4jXW4GDetOcrBP0H4zjiB49WLQIeBlaPuNm6oc6CQZQGgIOqOcwyLhx90eEC81Nxn1UWfdIH/0829Q7EDJqsAgcTl7cTZr+OlOM1gnqKzoEWR9xUIHZ31E+emOnPQC50iyUjD4wYsjNnWluViNrolU4LMLwS9dHPAgMBAAECggEALGEW93iH+p+O9V0Fdh1uL1OMM0ZczypNpr+q138WB9oRn4dcF5CxdXY+1nv3NGktwiNRUlpv0OJTDUF/TIDEaMoitw5Dz1cMEJoDC8+EvL4C3y5/Pq0AgVMnaWSRnNr/PpI3K5bo86bKRJ6a3cwIZjhPiY+1VU3+zmmtF5LdNUUGP9CFSu9cOkyQ6XARGSyCHUu6cx8qnuDAQhbvFKia9Wds7wmO+BmZKMe5h1AX52Qazm5FbzfyohoV+O/DeBFydcyXwMxDfQq/qOOJ2pLX31zjgC8xyfWHimAE1fzVYK0gCIuQTw6idVxP09Dok352M0uKxQB8Y7vjo0NndpqhCQKBgQC1dASBl7bQbU7X0iHpqE0/lgYhBoJ5p+NET3gas/UReDvRfVNgkJi2SQUB36jhsFC4Ss2eMdpJu3PgAsOOq2ukwUIu6H7/3iDyyBRj28efnYtVbGqxnbxOLyXb2lwQM3y+l49IgtTC4I0+sywSz2owsYt++Q/1LLYmxNkydLzAYwKBgQC2Ere+3l4S1n0Z3ShlAEtsiiY+iTEQfVaFfkgdrm4ZzVfcg1dKHCLfDbOj4Sb6h4tUvq0GUArxZw6MLmPw1A2CR6MRUnLZzF2imxqbxD1C1Ziw/3YFLzBUL6SYUvIPFIDQkAtV6ZiZ0i97AFKPsXSE3ZFc5QrqBm39Z0WAnx6GpQKBgQClNvXDaXgzXtOi/tmR/oQXmWoLN/ok+xCbaiBZb/F77pXIsOvxyLbDgsjkY9lJZViZ9MIzgADZkax4EwwsIzjn0m+IdnjsUDx/0l1+bubcFn5/US/0eUKzjYpTHAcnWToD78uuojFHH1578Hf4yJJPs9jny68T9ZoEwsdggZo7MQKBgQCJCfeQyq/Fv/iy8dmIg7tBO43jfcEXJy9rFIxVAsfN1ddumtClcB3zAs4V5849vK1rIIT8flxuK+IwMm54gjKUvzSGQic9tOs5C0LwwLhXSIs/lVeQOV9HzgA1wiB8pzL5R6GwjTyVcjTYgHqjuq33nwjdAWvAjwxUMVmHoLwKHQKBgQCIXF2VFTnJR5H2koS8VvsVX4TK+HDQZJTURKHGTO0vmcJ2WrovF7x+clkGeSkE8HLnW5sTtiF3FyPukBaFml4/Y2Zlr4qbQrwWobyQZm2Fp1Gr+e+zKxaXtvnrVQbIPa/rPixy0Ao3SnXKijxHmFxkfCxqtflUbGseRcjH+Xkk0g==
+auth.jwt.accessToken.expirationTime=60
