+++ user-manager-service/src/test/java/com/crmfoodestablishment/user_auth_service/UserManagerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.crmfoodestablishment.user_auth_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserManagerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
