--- user-manager-service/src/test/java/com/crmfoodestablishment/usermanager/UserManagerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.crmfoodestablishment.usermanager;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserManagerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
